pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Eoin H", email = "UltimateHurl@gmail.com"},
]
dependencies = [
    "pandas>=1.5.1",
    "tweepy>=4.12.1",
    "pyyaml>=6.0",
]
requires-python = ">=3.8"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
collect = "python collect.py"
[tool.pdm]
